Wil: - Arrumei o bug da direção da animação de ataque do herói estar errada depois de mudar a resolução. - Novas imagens das múmias e do faraó!!! - Retirei um código redundante da classe Floor.
git-svn-id: http://xp-dev.com/svn/horus_eye/trunk@474 863e7ce6-8861-447e-8dcb-03987b876380

+++ src/game/sprites/creature.cc
@@ -215,7 +215,7 @@ void Creature::HandleCollision(WorldObject* obj) {
     obj->CollidesWith(this);
 }
 
-int Creature::GetAttackingAnimationIndex(double angle) {
+int Creature::GetAttackingAnimationIndex(float angle) {
     int degreeAngle = (int)((angle / PI) * 360);
     degreeAngle += 45;
     int animationIndex = degreeAngle / 90;
